v5.9.0

+++ bus-core/src/main/java/org/aoju/bus/core/image/Image.java
@@ -50,7 +50,7 @@ import java.nio.file.Path;
  * 图像编辑器
  *
  * @author Kimi Liu
- * @version 5.8.9
+ * @version 5.9.0
  * @since JDK 1.8+
  */
 public class Image implements Serializable {
@@ -262,7 +262,7 @@ public class Image implements Serializable {
 
         // PNG图片特殊处理
         if (FileType.TYPE_PNG.equals(this.targetImageType)) {
-            final AffineTransformOp op = new AffineTransformOp(AffineTransform.getScaleInstance((double) scale, (double) scale), null);
+            final AffineTransformOp op = new AffineTransformOp(AffineTransform.getScaleInstance(scale, scale), null);
             this.targetImage = op.filter(ImageUtils.toBufferedImage(srcImage), null);
         } else {
             final String scaleStr = Float.toString(scale);
@@ -543,7 +543,6 @@ public class Image implements Serializable {
      *
      * @param degree 旋转角度
      * @return 旋转后的图片
-     * @since 5.8.9
      */
     public Image rotate(int degree) {
         final java.awt.Image image = getValidSrcImg();
